// projects/steam-generation-assessment-lib/src/app/interfaces/proposed-setup.interface.ts
export interface ProposedSetupInterface {
  benchmarkBoilerEfficiency: number;
  benchmarkCondensateReturn: number;
  benchmarkCondensateReturnedPercentage: number;
  benchmarkCondensateTemperature: number;
  benchmarkDsiPressure: number;
  benchmarkTemperatureOfFeedtank: number;
  benchmarkWaterRejectionRate: number;
  condensateReturnUnit: number;
  condensateTemperatureUnit: number;
  dsiPressureUnit: number;
  economiserRequired: boolean;
  proposalBoilerEfficiency: number;
  proposalCondensateReturned: number;
  proposalCondensateReturnedPercentage: number;
  proposalCondensateTemperature: number;
  proposalCostOfSodiumSulphite: number;
  proposalDsiPressure: number;
  proposalTemperatureOfFeedtank: number;
  proposalWaterRejectionRate: number;
  temperatureOfFeedtankUnit: number;
}